quote-comparison-chatbot/requirements.txt
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.7
langchain-openai==0.3.5
langchain-text-splitters==0.3.2

chromadb==0.4.24
sentence-transformers==2.6.1

streamlit==1.38.0
python-dotenv==1.0.1
pdfplumber==0.10.3
